// src/enums.ts
export enum FormEnum {
  scalar = 'scalar',
  array = 'array',
  map = 'map',
  object = 'object',
  set = 'set',
  function = 'function',
  container = 'container',
  void = 'void',
}

export enum TypeofEnum {
  undefined = "undefined",
  object = "object",
  boolean = "boolean",
  number = "number",
  bigint = "bigint",
  string = "string",
  symbol = "symbol",
  function = "function"
}

export enum NumberEnum {
  infinite = "infinite",
  nan = "nan",
  integer = "integer",
  decimal = "decimal",
}

export enum TypeEnum {
  string = 'string',
  number = 'number',
  boolean = 'boolean',
  symbol = 'symbol',
  array = 'array',
  map = 'map',
  object = 'object',
  set = 'set',
  null = 'null',
  undefined = 'undefined',
  function = 'function'
}